refactor: convert perfoverlay into feature
