+++ features/Performance Overlay/Shaders/Features/PerformanceOverlay.ini
@@ -1,0 +1,2 @@
+[Info]
+Version = 1-0-0
