+++ src/Globals.h
@@ -24,6 +24,7 @@ struct VolumetricLighting;
 struct VR;
 struct WaterEffects;
 struct WeatherPicker;
+struct PerformanceOverlay;
 struct WetnessEffects;
 struct ExtendedTranslucency;
 
@@ -78,6 +79,7 @@ namespace globals
 		extern VR* vr;
 		extern WaterEffects* waterEffects;
 		extern WeatherPicker* weatherPicker;
+		extern PerformanceOverlay* performanceOverlay;
 		extern WetnessEffects* wetnessEffects;
 		extern ExtendedTranslucency* extendedTranslucency;
