+++ src/Menu.h
@@ -24,7 +24,6 @@ public:
 	void Init();
 	void DrawSettings();
 	void DrawOverlay();
-	void DrawPerfOverlay();
 	void DrawWeatherDetailsWindow();
 
 	void ProcessInputEvents(RE::InputEvent* const* a_events);
@@ -166,90 +165,19 @@ public:
 		uint32_t ToggleKey = VK_END;
 		uint32_t SkipCompilationKey = VK_ESCAPE;
 		uint32_t EffectToggleKey = VK_MULTIPLY;  // toggle all effects
+		uint32_t OverlayToggleKey = VK_F10;      // Global overlay toggle key for all overlays
 		ThemeSettings Theme;
-
-		struct PerfOverlaySettings
-		{
-			bool Enabled = false;
-			bool ShowDrawCalls = true;
-			bool ShowVRAM = true;
-			bool ShowFPS = true;
-			bool ShowPreFGFrameTimeGraph = true;
-			bool ShowPostFGFrameTimeGraph = true;
-			float UpdateInterval = 0.5f;
-			int FrameHistorySize = 120;                       // Default 120 frames = 2s @ 60fps. Clamped using static values to prevent config file values going outside of slider bounds.
-			static constexpr int kMinFrameHistorySize = 60;   // 60 frames = 1s @ 60fps. Reasonable minimum.
-			static constexpr int kMaxFrameHistorySize = 480;  // 480 frames = 10s @ 60fps or 2s @ 240fps. Reasonable maximum.
-			enum class TextSize
-			{
-				Small,
-				Medium,
-				Large
-			};
-			TextSize Size = TextSize::Medium;
-
-			float BackgroundOpacity = 0.5f;
-			bool ShowBorder = true;
-			ImVec2 Position = ImVec2(10.f, 10.f);
-			bool PositionSet = false;
-			uint32_t OverlayToggleKey = VK_F10;
-		} PerfOverlay;
-
-		struct WeatherDetailsWindowSettings
-		{
-			bool Enabled = false;
-			ImVec2 Position = ImVec2(50.f, 50.f);
-			bool PositionSet = false;
-		} WeatherDetailsWindow;
 	};
-	const ThemeSettings& GetTheme() const { return settings.Theme; }  // Provide read-only access to the Theme.
-	Settings& GetSettings() { return settings; }                      // Provide access to settings for other components
+	const ThemeSettings& GetTheme() const { return settings.Theme; }                // Provide read-only access to the Theme.
+	Settings& GetSettings() { return settings; }                                    // Provide access to settings for other components
+	winrt::com_ptr<IDXGIAdapter3> GetDXGIAdapter3() const { return dxgiAdapter3; }  // Provide access to dxgiAdapter3
 
 	void SelectFeatureMenu(const std::string& featureName);
 	static std::unordered_map<std::string, int> categoryCounts;  // Number of features in each feature category
 
-	class PerfOverlayState
-	{
-	public:
-		std::vector<float> frameTimeHistory;
-		std::vector<float> postFGFrameTimeHistory;
-		bool initialized = false;
-		bool hasGraphs = false;
-		int frameTimeHistoryIndex = 0;
-		int postFGFrameTimeHistoryIndex = 0;
-		bool isFrameGenerationActive = false;
-		int64_t frequency;
-		int64_t lastFrameCounter;
-		int64_t currentFrameCounter;
-		float frameTimeMs = 0.0f;
-		float fps = 0.0f;
-		float postFGFrameTimeMs = 0.0f;
-		float postFGFps = 0.0f;
-		float smoothFps = 0.0f;
-		float smoothFrameTimeMs = 0.0f;
-		float postFGSmoothFps = 0.0f;
-		float postFGSmoothFrameTimeMs = 0.0f;
-		float updateTimer = 0.0f;
-		float minFrameTime = 1000.0f;
-		float maxFrameTime = 0.0f;
-		float smoothedMinFrameTime = 0.0f;
-		float smoothedMaxFrameTime = 50.0f;
-		float textScale = 1.0f;
-		static constexpr float kSmoothingFactor = 0.15f;  // Smoothing factor: 0.1f = slow, 0.3f = fast.
-		std::chrono::steady_clock::time_point lastUpdateTime;
-		float SetTextScale(Settings::PerfOverlaySettings& settings);
-		void UpdateGraphValues(Settings::PerfOverlaySettings& settings);
-		void UpdateFrameTimeHistorySizes(Settings::PerfOverlaySettings& settings);
-		void UpdateMinFrameTime();
-		void UpdateMaxFrameTime();
-		void UpdateFGFrameTime(Settings::PerfOverlaySettings& settings);
-		void DrawPostFGFrameTimeGraph(Settings::PerfOverlaySettings& settings);
-		void DrawDrawCalls();
-		void DrawFPS(Settings::PerfOverlaySettings& settings);
-		void DrawVRAM(winrt::com_ptr<IDXGIAdapter3> dxgiAdapter3);
-	};
 	bool usingTestConfig = false;
-	PerfOverlayState perfOverlayState;
+	bool abTestingEnabled = false;
+	bool overlayVisible = false;
 
 	// Static utility functions
 	static const char* KeyIdToString(uint32_t key);
@@ -281,7 +209,6 @@ private:
 	void DrawDisplaySettings();
 	void DrawDisableAtBootSettings();
 	void DrawFooter();
-	void DrawPerformanceOverlaySettings();
 	void BuildCategoryCounts();
 
 	class CharEvent : public RE::InputEvent
